fix(bench): SetWriteBufferSize takes uint64 in grocksdb v1.8.12
Both RocksDB WAL test programs passed an int, so `go build ./...` failed on
them. Neither is on any experiment path, which is why it went unnoticed — but
it breaks a whole-repo build, and anything that breaks the build gets in the
way of the next person running the tests.

+++ benchmark/testRocksDB_v2/testRocksDB-v2.go
@@ -64,7 +64,7 @@ func (wt *WALTest) setupDBOptions() {
 	wt.Options.SetCreateIfMissing(true)
 	
 	// 调整写入缓冲区，避免过早flush到SST，让数据保留在WAL中
-	wt.Options.SetWriteBufferSize(int(wt.TargetSize * 2)) // 设置为目标WAL大小的2倍
+	wt.Options.SetWriteBufferSize(uint64(wt.TargetSize * 2)) // 设置为目标WAL大小的2倍
 	wt.Options.SetMaxWriteBufferNumber(3)
 	wt.Options.SetMinWriteBufferNumberToMerge(1)
